Move extra public headers to Internal and fix imports
