+++ GoogleDataTransport/GDTCCTTests/Common/TestStorage/GDTCCTTestStorage.h
@@ -17,7 +17,7 @@
 #import <Foundation/Foundation.h>
 #import <XCTest/XCTest.h>
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORStorageProtocol.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORStorageProtocol.h"
 
 NS_ASSUME_NONNULL_BEGIN
