+++ GoogleDataTransport/GDTCCTTests/Unit/TestServer/GDTCCTTestServer.m
@@ -16,7 +16,7 @@
 
 #import "GoogleDataTransport/GDTCCTTests/Unit/TestServer/GDTCCTTestServer.h"
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORAssert.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORAssert.h"
 
 #import <nanopb/pb_decode.h>
 #import <nanopb/pb_encode.h>
