+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORDirectorySizeTracker.h
@@ -16,7 +16,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORStorageProtocol.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORStorageProtocol.h"
 
 NS_ASSUME_NONNULL_BEGIN
