+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORPlatform.h
