+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORRegistrar.h
