+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORStorageEventSelector.h
