+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORStorageProtocol.h
