+++ GoogleDataTransport/GDTCORLibrary/Internal/GDTCORUploader.h
