+++ GoogleDataTransport/GDTCORLibrary/Private/GDTCORReachability_Private.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORReachability.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORReachability.h"
 
 @interface GDTCORReachability ()
