+++ GoogleDataTransport/GDTCORLibrary/Private/GDTCORRegistrar_Private.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORRegistrar.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORRegistrar.h"
 
 @interface GDTCORRegistrar ()
