+++ GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GoogleDataTransport.h
@@ -19,8 +19,5 @@
 #import "GDTCOREvent.h"
 #import "GDTCOREventDataObject.h"
 #import "GDTCOREventTransformer.h"
-#import "GDTCORLifecycle.h"
-#import "GDTCORRegistrar.h"
 #import "GDTCORTargets.h"
 #import "GDTCORTransport.h"
-#import "GDTCORUploader.h"
