+++ GoogleDataTransport/GDTCORTests/Common/Categories/GDTCORRegistrar+Testing.h
@@ -15,7 +15,7 @@
  */
 
 #import "GoogleDataTransport/GDTCORLibrary/Private/GDTCORRegistrar_Private.h"
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORRegistrar.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORRegistrar.h"
 
 NS_ASSUME_NONNULL_BEGIN
