+++ GoogleDataTransport/GDTCORTests/Common/Categories/GDTCORUploadCoordinator+Testing.m
@@ -19,7 +19,7 @@
 #import <objc/runtime.h>
 
 #import "GoogleDataTransport/GDTCORLibrary/Private/GDTCORFlatFileStorage.h"
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORRegistrar.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORRegistrar.h"
 
 @implementation GDTCORUploadCoordinator (Testing)
