+++ GoogleDataTransport/GDTCORTests/Common/Fakes/GDTCORApplicationFake.h
@@ -16,7 +16,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORPlatform.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORPlatform.h"
 
 NS_ASSUME_NONNULL_BEGIN
