+++ GoogleDataTransport/GDTCORTests/Common/Fakes/GDTCORStorageFake.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#import "GoogleDataTransport/GDTCORLibrary/Public/GoogleDataTransport/GDTCORStorageProtocol.h"
+#import "GoogleDataTransport/GDTCORLibrary/Internal/GDTCORStorageProtocol.h"
 
 NS_ASSUME_NONNULL_BEGIN
